+++ common/src/sema4ai/common/autoexit.py
@@ -11,9 +11,7 @@ _watching_thread_global = None
 PARENT_PROCESS_WATCH_INTERVAL = 2
 
 
-def exit_when_pid_exists(
-    pid: str | int, interval=PARENT_PROCESS_WATCH_INTERVAL
-) -> None:
+def exit_when_pid_exits(pid: str | int, interval=PARENT_PROCESS_WATCH_INTERVAL) -> None:
     """
     Exit the current process when the given pid exists (may be called multiple times and if
     any of the pids we're tracking exit, the current process will exit).
@@ -52,47 +50,18 @@ def exit_when_pid_exists(
             _watching_thread_global.start()
 
 
+# Keep old API (which has a typo).
+exit_when_pid_exists = exit_when_pid_exits
+
+
 def _os_exit(retcode: int) -> None:
     """
     Kills subprocesses and exits with the given returncode.
     """
+    from sema4ai.common.process import kill_subprocesses
+
     try:
-        import psutil
-
-        curr_process = psutil.Process()
-        try:
-            try:
-                children_processes = list(curr_process.children(recursive=True))
-            except Exception:
-                # Retry once
-                children_processes = list(curr_process.children(recursive=True))
-
-            try:
-                names = ",".join(f"{x.name()} (x.pid)" for x in children_processes)
-            except Exception as e:
-                log.debug(f"Exception when collecting process names: {e}")
-                names = "<unable to get>"
-
-            log.info(f"sema4ai-action-server killing processes after run: {names}")
-            for p in children_processes:
-                try:
-                    p.kill()
-                except Exception as e:
-                    log.debug(f"Exception when terminating process: {p.pid}: {e}")
-
-            # Give processes 2 seconds to exit cleanly and force-kill afterwards
-            _gone, alive = psutil.wait_procs(children_processes, 2)
-            for p in alive:
-                try:
-                    p.terminate()
-                except Exception as e:
-                    # Expected: process no longer exists.
-                    log.debug(f"Exception when killing process: {p.pid}: {e}")
-            # Wait a bit more after terminate.
-            psutil.wait_procs(alive, 2)
-        except Exception as e:
-            log.debug(f"Exception when listing/killing processes: {e}")
-
+        kill_subprocesses()
         sys.stdout.flush()
         sys.stderr.flush()
         # Give some time for other threads to run just a little bit more.